+++ threedigrid/admin/gridadmin.py
@@ -131,7 +131,7 @@ class GridH5Admin:
     @property
     def breaches(self):
         return Breaches(
-            self.datasource_class(self.h5py_file, "breaches"), **self._grid_kwargs
+            self.datasource_class(self.h5py_file, "breaches", gridadmin=self), **self._grid_kwargs
         )
 
     @property